src/requirements.txt
auto_mix_prep==0.2.0
Flask==3.1.0
Flask_JWT_Extended==4.7.1
Flask_MySQLdb==2.0.0
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
flask-swagger-ui==4.11.1
Flask-Cors==4.0.0
